pyproject.toml
[tool.poetry]
name = "target-google-ads"
version = "0.1.0"
description = ""
authors = ["Deke Li <deke.li@sendinblue.com>"]

[tool.poetry.dependencies]
python = "^3.9"
singer-python = "^5.12.2"
google-ads = "^16.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
